+++ environ/environ.py
@@ -154,8 +154,9 @@ class Env:
         "simple": "haystack.backends.simple_backend.SimpleEngine",
     }
 
-    def __init__(self, **scheme):
+    def __init__(self, interpolate=False, **scheme):
         self.smart_cast = True
+        self.interpolate = interpolate
         self.scheme = scheme
 
     def __call__(self, var, cast=None, default=NOTSET, parse_default=False):
@@ -347,7 +348,8 @@ class Env:
             value = default
 
         # Resolve any proxied values
-        if hasattr(value, 'startswith') and value.startswith('$'):
+        if self.interpolate and hasattr(value, 'startswith') and \
+                value.startswith('$'):
             value = value.lstrip('$')
             value = self.get_value(value, cast=cast, default=default)