+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Breaking Changes
   ``Env.read_env()`` will assume there's no ``.env`` file to be found, log a
   WARN-level log message to that effect, and continue on.
 * Make ``Env.read_env()``'s ``overrides`` argument actually override variables.
+* Disabled proxy variable feature by default.
 
 
 Features