+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 This file contains a brief summary of new features and dependency changes or
 releases, in reverse chronological order.
 
-2.1.0 (2021-XX-XX)
+2.1.0 (2021-05-29)
 ------------------